pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]

[project]
dependencies = [
  "dnspython>=2.6.1",
  "aiohttp>=3.9.3",
  "typer>=0.9.0",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1",
  "beautifulsoup4>=4.12.3",
  "certifi>=2024.2.2",
  "typing-extensions>=4.10.0",
  "cachetools>=5.3.3",
  "backoff>=2.2.1",
  "pysocks>=1.7.1",
  "requests[socks]>=2.31.0",
  "aiofiles>=23.2.1",
]
description = "Ferramenta modular de pentest DNS"
name = "cataclysmdns"
requires-python = ">=3.10"
version = "0.2.0"

[[project.authors]]
name = "Daniel Petronilha"

[project.scripts]
cataclysm = "cli:app"

[tool.setuptools]
packages = ["core", "utils"]
py-modules = ["cli", "config"]
